Structure mostly fixed
Wow, this took me 9 days? Awful. December is not a good month for productivity.

The structure is mostly fixed now. I genuinely have no clue why I was in such a crisis before, but with this new format I have less repetition and less files to thumb through. Its much much clearer.

I'm going to delay programming any more devices for as long as possible because I don't know how to make which files show on the solution explorer more selective. That's the actual biggest reason.

Every time I use [] on a map, to check if it contains that key already, I think it creates data corruption?? I really don't understand why, because the program is able to check that the key isn't present and leave that operation safely, and then it's the next time the map is used that it turns out its like all corrupted? In any other language, that would work fine with a dictionary. I'm curious what is different about the map in C++, or what kind of protections are missing.