Sleep Timers and Schedules
Having them both be schedules in this way is messy. It started like this because I knew that, functionally, they work exactly the same. And they do! But the inputs from the user to make them are slightly different, and the text that needs to be displayed is sort of different! So this schedule should really be the base class and I should've made subclasses for those parts. I might do that later, but probably not.

I think that I'm, supposed to be making interfaces for every class, classes inheriting them just to implement the controls, and then CL_Menu would be juggling those instead. I'm not because I just can't handle having to flip through that many files in VS2022. If it were VC Code and I knew how to get the compiler to acknowledge files from different subfolders then it would be fine. Alas.